--- README.md
@@ -1,1 +1,0 @@
-Yelp Chalenge Adjusted Rating App
