--- process.py
@@ -1,23 +1,0 @@
-import dash
-import dash_bootstrap_components as dbc
-import dash_core_components as dcc
-import dash_html_components as html
-from dash.dependencies import Input, Output
-
-from app import app
-
-column1 = dbc.Col(
-    [
-        dcc.Markdown(
-            """
-        
-            ## Process
-
-
-            """
-        ),
-
-    ],
-)
-
-layout = dbc.Row([column1])
