--- pull_request_template.md
@@ -1,35 +1,0 @@
-# Description
-
-Fixes # (issue)
-
-## Type of change
-
-Please delete options that are not relevant.
-
-- [ ] Bug fix (non-breaking change which fixes an issue)
-- [ ] New feature (non-breaking change which adds functionality)
-- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
-- [ ] This change requires a documentation update
-
-## Change Status
-
-- [ ] Complete, tested, ready to review and merge
-- [ ] Complete, but not tested (may need new tests)
-- [ ] Incomplete/work-in-progress, PR is for discussion/feedback
-
-# How Has This Been Tested?
-
-- [ ] Test A
-- [ ] Test B
-
-# Checklist
-
-- [ ] My code follows the style guidelines of this project
-- [ ] I have performed a self-review of my own code
-- [ ] My code has been reviewed by at least one peer
-- [ ] I have commented my code, particularly in hard-to-understand areas
-- [ ] I have made corresponding changes to the documentation
-- [ ] My changes generate no new warnings
-- [ ] I have added tests that prove my fix is effective or that my feature works
-- [ ] New and existing unit tests pass locally with my changes
-- [ ] There are no merge conflicts
