+++ yelp-challenge-adjusted-rating-app/Pipfile
@@ -1,0 +1,19 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+dash = "*"
+dash-bootstrap-components = "*"
+gunicorn = "*"
+plotly = "*"
+pandas = "*"
+joblib = "*"
+scikit-learn = "*"
+category-encoders = "*"
+
+[requires]
+python_version = "3"
