+++ yelp-challenge-adjusted-rating-app/Procfile
@@ -1,0 +1,1 @@
+web: gunicorn run:server
